Make the pipeline depth an int instead of a size_t. While they are supposed to be unsigned, casting them to a signed value for all atomic operations is as errorprone as handling them as signed entities.

+++ ompi/mca/pml/ob1/pml_ob1.h
@@ -55,8 +55,8 @@ struct mca_pml_ob1_t {
     int free_list_num;      /* initial size of free list */
     int free_list_max;      /* maximum size of free list */
     int free_list_inc;      /* number of elements to grow free list */
-    size_t send_pipeline_depth;
-    size_t recv_pipeline_depth;
+    int32_t send_pipeline_depth;
+    int32_t recv_pipeline_depth;
     size_t rdma_retries_limit;
     int max_rdma_per_request;
     int max_send_per_range;
